+++ src/Api/Discovery/CannotDisableBindingException.php
@@ -75,20 +75,4 @@ class CannotDisableBindingException extends RuntimeException
     {
         return static::forUuid($uuid, $packageName, 'The type of the binding is not loaded.', $code, $cause);
     }
-
-    /**
-     * Creates an exception for a UUID that could not be disabled because the
-     * binding type is defined twice.
-     *
-     * @param Uuid      $uuid        The UUID.
-     * @param string    $packageName The name of the package.
-     * @param int       $code        The exception code.
-     * @param Exception $cause       The exception that caused this exception.
-     *
-     * @return static The created exception.
-     */
-    public static function duplicateType(Uuid $uuid, $packageName, $code = 0, Exception $cause = null)
-    {
-        return static::forUuid($uuid, $packageName, 'The type definition is duplicated.', $code, $cause);
-    }
 }